tests/wpt/tests/html/canvas/element/canvas-context/2d.canvas.context.exists.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.canvas.context.exists</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.canvas.context.exists</h1>
<p class="desc">The 2D context is implemented</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("The 2D context is implemented");
_addTest(function(canvas, ctx) {

  _assertDifferent(canvas.getContext('2d'), null, "canvas.getContext('2d')", "null");

});
</script>
